% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_article.R
\name{CREDIT_to_CSV}
\alias{CREDIT_to_CSV}
\title{Convert the 'xlsx' file with the credit information to csv format.}
\usage{
CREDIT_to_CSV(art_path, art_id, verbose)
}
\arguments{
\item{art_path}{A string with the path to the article folder.}

\item{art_id}{A string with the article id.}

\item{verbose}{Logical. Specifies whether to display verbose output (default
\code{TRUE}).}
}
\value{
Invisible TRUE.
}
\description{
Convert the 'xlsx' file with the credit information to csv format.
}
