% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri-package.R
\docType{package}
\name{guri-package}
\alias{guri}
\alias{guri-package}
\title{guri: ~!gurí_: Unified Format Manager for Research Journals}
\description{
~gurí_ (Gestor Unificado de formatos para Revistas de Investigación / Unified Format Manager for Research Journals) facilitates the generation of final documents for scientific journals from documents obtained in the 'proofreading' stage. The proposal seeks to solve the difficulties of some academic journals in generating final documents in different formats in a consistent way and without generating duplicated processes. It also takes into account that many scientific journals use docx documents as the basis of their workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/estedeahora/guri}
  \item \url{https://estedeahora.github.io/guri/}
  \item Report bugs at \url{https://github.com/estedeahora/guri/issues}
}

}
\author{
\strong{Maintainer}: Pablo Santiago Serrati \email{pabloserrati@gmail.com} (\href{https://orcid.org/0000-0001-5300-2243}{ORCID})

}
\keyword{internal}
