% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri.R
\name{guri}
\alias{guri}
\title{Función para generar archivos finales de un artículo individual}
\usage{
guri(
  art_path,
  art_name,
  journal = NULL,
  verbose = F,
  zip_file = F,
  clean_files = T
)
}
\arguments{
\item{art_path}{String...}

\item{art_name}{String...}

\item{journal}{String... description}

\item{verbose}{Logical...}

\item{zip_file}{Logical...}

\item{clean_files}{Logical...}
}
\value{
Invisible. ...
}
\description{
describir
}
