% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_appendix.R
\name{guri_appendix}
\alias{guri_appendix}
\title{Make markdown appendix from docx documents}
\usage{
guri_appendix(art_path, art_id)
}
\arguments{
\item{art_path}{A string with the path to the article folder, where the appendix files are located.}

\item{art_id}{A string with the article id, used to identify the appendix files. The appendix files should be named as "art_app1.docx", "art_app2.docx", etc.}
}
\value{
A character vector containing the names of the converted Markdown files.
}
\description{
Make markdown appendix from docx documents
}
