% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_outputs.R
\name{guri_clean_files}
\alias{guri_clean_files}
\title{Clean temporary log and output files}
\usage{
guri_clean_files(path, art_id, verbose)
}
\arguments{
\item{path}{A string with the path to the article folder.}
}
\value{
Invisible TRUE.
}
\description{
Clean temporary log and output files
}
