% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_config_journal.R
\name{guri_config_journal}
\alias{guri_config_journal}
\title{Create the journal configuration files}
\usage{
guri_config_journal(journal_folder = NULL, csl_name = NULL, force = FALSE)
}
\arguments{
\item{journal_folder}{A string with the path to the journal. If \code{NULL} (default),
the working folder is used. For journal repositories, this must be provided.}

\item{csl_name}{A string with the CSL's name (without its extension).
See: https://github.com/citation-style-language/styles}

\item{force}{Logical. Should it be overwritten if a configuration folder already
exists (Default: FALSE)?}
}
\value{
Invisible \code{TRUE}.
}
\description{
Create the journal configuration files in \verb{_config}. If \code{csl_name} is given,
it also places the corresponding csl in the configuration folder.
}
