% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_convert.R
\name{guri_convert}
\alias{guri_convert}
\title{Converts the corrected manuscript between the formats required by \code{~!guri_}.}
\usage{
guri_convert(path_art, art, output, verbose = F)
}
\arguments{
\item{path_art}{A string with the path to the article directory}

\item{art}{A string with the article id.}

\item{output}{String}

\item{verbose}{Logical}
}
\value{
Invisible
}
\description{
Converts the corrected manuscript between the formats required by \code{~!guri_}.
}
