% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guri_CREDITtoCSV}
\alias{guri_CREDITtoCSV}
\title{TODO Prepare the working directory for each article.}
\usage{
guri_CREDITtoCSV(art_path, art_pre, verbose = F)
}
\arguments{
\item{art_path}{A string with the path to the article folder.}

\item{art_pre}{A string. This prefix is used to identify the folders where each issue of your journal will be stored. For example, if you use 'num' (default) the folders where you should store the issues of your journal will be 'num1', 'num2', and so on.}

\item{verbose}{Logical}
}
\value{
Invisible True if succesfull.
}
\description{
TODO Prepare the working directory for each article.
}
