% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_doi_batch.R
\name{guri_doi_batch}
\alias{guri_doi_batch}
\title{Generate a \code{doi_batch} xml file for Crossref deposit}
\usage{
guri_doi_batch(list_art, path_issue)
}
\arguments{
\item{list_art}{A data.frame with the article information as returned by
\link{guri_list_articles}. It must contain the following columns: \code{id}
(character), with the article id; and \code{path} (character), with the path to
the article.}

\item{path_issue}{A string with the path to the issue folder.}
}
\value{
A string with the path to the \code{doi_batch} xml file.
}
\description{
Generate a \code{doi_batch} xml file for Crossref deposit
}
\details{
The \code{doi_batch} xml file is saved in the \code{doi_register} folder,
inside the issue folder.

The \code{doi_batch} xml file searches each of the article folders for the
\verb{art[id]_crossref.xml} file (inside the '_temp' folder), and uses the
information in these files to create a unified \code{doi_batch} xml file (each
article is saved as a \verb{<journal_article>} element in the xml file).

In adition to the \code{doi_batch} xml file, this function also creates a text
file with the information of the articles that will be deposited in Crossref.
The text file is saved in the same folder as the \code{doi_batch} xml file.
}
