% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_install.R
\name{guri_install}
\alias{guri_install}
\title{Updates/installs the external dependencies necessary for working \code{~!guri_}.}
\usage{
guri_install(pandoc = T, tinytex = T, force = F)
}
\arguments{
\item{pandoc}{Logical. Should pandoc be installed/updated? (default 'TRUE')}

\item{tinytex}{Logical. Should tinytex be installed/updated? (default 'TRUE')}

\item{force}{Logical. Should pandoc  and tinytex be forced to reinstall? (default 'FALSE')}
}
\value{
Invisible. A list with a logical vector indicating whether tinytex
and pandoc are available and two items with the installed versions of Tinitex and Pandoc.
}
\description{
Upgrade or install (as necessary) pandoc, as well as the latex distribution
(tinytex) and latex packages.
}
