% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_list_articles.R
\name{guri_list_articles}
\alias{guri_list_articles}
\title{Lists the articles in a journal issue}
\usage{
guri_list_articles(path_issue)
}
\arguments{
\item{path_issue}{String with the path to the issue folder to list the
articles in the issue.}
}
\value{
A tibble with two columns: articles path (art_path) and articles id (art_id).
}
\description{
Lists the articles in a journal issue
}
