% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_outputs.R
\name{guri_outputs}
\alias{guri_outputs}
\title{Función para generar archivos finales}
\usage{
guri_outputs(
  art_id,
  issue,
  journal = NULL,
  doi_batch = FALSE,
  verbose = FALSE,
  clean_files = TRUE
)
}
\arguments{
\item{art_id}{String...}

\item{issue}{String...}

\item{journal}{String... description}

\item{doi_batch}{Logical...}

\item{verbose}{Logical...}

\item{clean_files}{Logical...}
}
\value{
Invisible. ...
}
\description{
describir
}
