% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri.R
\name{CREDIT_to_CSV}
\alias{CREDIT_to_CSV}
\title{Convert the 'xlsx' file with the credit information to csv format.}
\usage{
CREDIT_to_CSV(path, art_id, verbose)
}
\arguments{
\item{path}{A string with the path to the article folder.}

\item{art_id}{A string. The 'article id'.}

\item{verbose}{Logical.}
}
\value{
Invisible TRUE.
}
\description{
Convert the 'xlsx' file with the credit information to csv format.
}
