% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_appendix.R
\name{guri_appendix}
\alias{guri_appendix}
\title{DESCRIBE}
\usage{
guri_appendix(wdir, art)
}
\arguments{
\item{wdir}{description}

\item{art}{description}
}
\value{
Returns the name of the files containing the appendices in markdown format.
}
\description{
DESCRIBE
}
