% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_appendix.R
\name{guri_appendix}
\alias{guri_appendix}
\title{Make markdown appendix from docx documents}
\usage{
guri_appendix(wdir, art)
}
\arguments{
\item{wdir}{The working directory where the appendix files are located.}

\item{art}{The article name used to identify the appendix files. The appendix files should be named as "art_app1.docx", "art_app2.docx", etc.}
}
\value{
A character vector containing the names of the converted Markdown files.
}
\description{
Make markdown appendix from docx documents
}
