% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_article.R
\name{guri_article}
\alias{guri_article}
\title{Generate the output files for individual article.}
\usage{
guri_article(art_path, art_dir, art_id, verbose = TRUE, clean_files = TRUE)
}
\arguments{
\item{art_path}{A string with the path to the article folder.}

\item{art_dir}{A string with the article folder name.}

\item{art_id}{A string with the article id.}

\item{verbose}{Logical. Specifies whether to display verbose output (default
\code{TRUE}).}

\item{clean_files}{Logical. Should the temporary files be deleted and
reordered in folders after the creation of the final files?. Primarily for
debugging purposes (default is \code{TRUE}).}
}
\value{
Invisible TRUE.
}
\description{
Generate the output files for individual article.
}
