% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri.R
\name{guri_article}
\alias{guri_article}
\title{Función para generar archivos finales de un artículo individual}
\usage{
guri_article(
  art_path,
  art_id,
  journal = NULL,
  verbose = FALSE,
  clean_files = TRUE
)
}
\arguments{
\item{art_path}{String...}

\item{art_id}{String...}

\item{journal}{String... description}

\item{verbose}{Logical...}

\item{clean_files}{Logical...}
}
\value{
Invisible. ...
}
\description{
describir
}
