% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_article.R
\name{guri_article}
\alias{guri_article}
\title{Función para generar archivos finales de un artículo individual}
\usage{
guri_article(art_path, art_dir, art_id, verbose = T, clean_files = T)
}
\arguments{
\item{art_path}{String... description}

\item{art_dir}{String... description}

\item{art_id}{String...}

\item{verbose}{Logical...}

\item{clean_files}{Logical...}
}
\value{
Invisible. ...
}
\description{
describir
}
