% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_clean_files.R
\name{guri_clean_files}
\alias{guri_clean_files}
\title{Clean and reorganize article folder, moving temporary (auxiliary), log and output files.}
\usage{
guri_clean_files(art_path, art_id, verbose)
}
\arguments{
\item{art_path}{A string with the path to the article folder.}

\item{art_id}{A string with the article id.}

\item{verbose}{Logical. Specifies whether to display verbose output (default
\code{TRUE}).}
}
\value{
Invisible TRUE.
}
\description{
Clean and reorganize article folder, moving temporary (auxiliary), log and output files.
}
\seealso{
\link{guri_outputs}
}
