% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_clean_files.R
\name{guri_clean_files}
\alias{guri_clean_files}
\title{Clean temporary log and output files}
\usage{
guri_clean_files(art_path, art_id, verbose)
}
\arguments{
\item{art_path}{A string with the path to the article folder.}

\item{art_id}{String or vector of strings with the article id to be processed
in the issue. If "all" all articles are processed.}

\item{verbose}{Logical. If \code{TRUE} the function will print more information
during the process (default is \code{FALSE}).}
}
\value{
Invisible TRUE.
}
\description{
Clean temporary log and output files
}
