% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_convert.R
\name{guri_convert}
\alias{guri_convert}
\title{Converts the corrected manuscript between the formats required by \code{~!guri_}.}
\usage{
guri_convert(path_art, art, output, verbose = F)
}
\arguments{
\item{path_art}{A string with the path to the article directory}

\item{art}{A string with the article id.}

\item{output}{A string. The desired output format (see ).}

\item{verbose}{Logical. Specifies whether to display verbose output.}
}
\value{
Invisible TRUE
}
\description{
This function converts a document using \code{~!guri_}. It takes the path to the
article, the article id, the desired output format, and an optional verbose
flag. It performs the conversion by calling the \link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert}
function.
}
