% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_to_format.R
\name{guri_to_md}
\alias{guri_to_md}
\alias{guri_to_html}
\alias{guri_to_jats}
\alias{guri_to_pdf}
\alias{guri_to_AST}
\alias{guri_biblio}
\title{Converts the corrected manuscript in docx format to markdown format.}
\usage{
guri_to_md(path_art, art, verbose = F)

guri_to_html(path_art, art, verbose = F)

guri_to_jats(path_art, art, verbose = F)

guri_to_pdf(path_art, art, verbose = F, pdf = TRUE)

guri_to_AST(path_art, art, verbose = F)

guri_biblio(path_art, art, verbose = F, bib_type = "csljson")
}
\arguments{
\item{path_art}{A string with the path to the article directory}

\item{art}{A string with the article id.}

\item{verbose}{Logical}

\item{pdf}{Logical. Should the pdf be generated? (default = TRUE)}

\item{bib_type}{description}
}
\value{
Invisible \code{TRUE}
}
\description{
They convert between the different formats required for the
\code{~!guri_} workflow, applying the appropriate lua filters at each stage and
using the appropriate templates and metadata. The \code{~!guri_to_md} function
converts between the revised (and formatted) manuscript in \code{docx} format to
markdown format. The \code{guri_to_html}, \code{guri_to_jats} and \code{guri_to_pdf}
functions convert from the markdown file to \code{html}, \code{xml} (xml-jats) and
\code{pdf} (\code{tex}) formats, respectively. Last, \code{guri_to_AST} and \code{guri_biblio}
are auxiliary functions that generate a file with the Abstract Syntax Tree
(mostly for debugging purposes) and a file with the references used by the
article, respectively.
}
\details{
These functions are, mostly, a wrapper of the internal function
\link{guri_convert}. The functions are exported primarily for debugging and error
detection purposes. For day-to-day work it is advisable to use the \link{guri}
function directly, which coordinates the generation of all these formats and
ensures the correct definition of the parameters.

For the generation of pdf files, LuaTex (\link[tinytex:latexmk]{tinytex::lualatex}) is used
internally.
}
