% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guri_to_output.R
\name{guri_to_md}
\alias{guri_to_md}
\alias{guri_to_html}
\alias{guri_to_jats}
\alias{guri_to_pdf}
\alias{guri_to_AST}
\alias{guri_biblio}
\title{Converts the corrected manuscript in docx format to markdown format.}
\usage{
guri_to_md(path_art, art, verbose = F)

guri_to_html(path_art, art, verbose = F)

guri_to_jats(path_art, art, verbose = F)

guri_to_pdf(path_art, art, verbose = F)

guri_to_AST(path_art, art, verbose = F)

guri_biblio(path_art, art, verbose = F, bib_type = "csljson")
}
\arguments{
\item{path_art}{A string with the path to the article directory}

\item{art}{A string with the article id.}

\item{verbose}{Logical}

\item{bib_type}{description}
}
\value{
Invisible
}
\description{
Converts the corrected manuscript in docx format to markdown format.

Genera archivo con bibliografia (en biblatex o csljson)

Genera archivo con bibliografia (en biblatex o csljson)
}
